+++ src/HttpClient/GenericHttpClient.php
@@ -24,12 +24,19 @@ class GenericHttpClient implements ReplaceableHttpClientInterface
     private $serializer;
     private $httpClient;
 
+    /**
+     * @var iterable<AuthenticationHeaderProviderInterface>
+     */
+    private $authenticationHeaderProviders;
+
     public function __construct(
         SerializerInterface $serializer,
-        HttpClientInterface $httpClient
+        HttpClientInterface $httpClient,
+        iterable $authenticationHeaderProviders = []
     ) {
         $this->serializer = $serializer;
         $this->httpClient = $httpClient;
+        $this->authenticationHeaderProviders = $authenticationHeaderProviders;
     }
 
     /**
@@ -55,6 +62,15 @@ class GenericHttpClient implements ReplaceableHttpClientInterface
 
         if (null !== $body) {
             $options['body'] = $this->serializer->serialize($body, $bodyFormat);
+        }
+
+        // Only the first authentication header provider that supports the context will add the authentication header.
+        foreach ($this->authenticationHeaderProviders as $provider) {
+            if ($provider->supports(['method' => $method, 'uri' => $uri, 'options' => $options, 'microservice' => $microservice])) {
+                $options['headers'] += [$provider->getHeader() => $provider->getValue()];
+
+                break;
+            }
         }
 
         return $this->httpClient->request($method, $uri, $options);